+++ tests/integration/test_credit_card_verification.py
@@ -169,3 +169,15 @@ class TestCreditCardVerfication(unittest.TestCase):
         self.assertEqual("XX", verification.network_response_code)
         self.assertEqual("sample network response text", verification.network_response_text)
 
+    def test_create_success_network_transaction_id(self):
+        result = CreditCardVerification.create({
+            "credit_card": {
+                "number": CreditCardNumbers.Visa,
+                "cardholder_name": "John Smith",
+                "expiration_date": "05/2012"
+            },
+        })
+
+        self.assertTrue(result.is_success)
+        verification = result.verification
+        self.assertRegex(verification.network_transaction_id, r'\d{15}')